fix extractor module move leftovers
