+++ tests/artificial/test_main.py
@@ -1,4 +1,4 @@
-from mex.pipeline import run_job_in_process
+from mex.extractors.pipeline import run_job_in_process
 
 
 def test_job() -> None:
