+++ tests/blueant/test_main.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mex.pipeline import run_job_in_process
+from mex.extractors.pipeline import run_job_in_process
 
 
 @pytest.mark.usefixtures(
