+++ tests/odk/test_main.py
@@ -1,7 +1,7 @@
 import pytest
 
 from mex.common.models import ExtractedPrimarySource
-from mex.pipeline import run_job_in_process
+from mex.extractors.pipeline import run_job_in_process
 
 
 @pytest.mark.usefixtures("mocked_ldap", "mocked_wikidata")
